requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
pytorch==1.3.1
argparse
transformers==3.0.0
sklearn
tqdm